lets check size read

+++ kernel_stack_grabber/kernel_stack_grabber.c
@@ -46,7 +46,7 @@ int before_block_exec(CPUState *env, TranslationBlock *tb) {
           	fp = fopen(str, "w+");
           	// char buf_str[10000];
           	// strcpy(buf_str, buf);
-          	printf("buf size: %d\n", sizeof(buf));
+          	printf("buf size: %d\n", (int)sizeof(buf));
           	fwrite(buf, 1, sizeof(buf), fp);
           	fclose(fp);
           	free(buf);
